requirements.txt
python-telegram-bot==13.13
PyYAML
google-auth==2.22.0
google-cloud-compute==1.15.0